refactor(cymbal): extract shared native symbolication module from apple
