+++ rust/cymbal/src/langs/apple.rs
@@ -12,11 +12,13 @@ use crate::{
     config::FRAME_CONTEXT_LINES,
     error::{AppleError, FrameError, ResolveError, UnhandledError},
     frames::{record_frame_resolution_failure, Frame},
+    langs::native::{self, DebugImage},
     langs::utils::{add_raw_to_junk, get_context_lines},
     langs::CommonFrameMetadata,
     symbol_store::{
-        apple::{AppleRef, ParsedAppleSymbols},
+        apple::AppleRef,
         chunk_id::OrChunkId,
+        native::{ParsedNativeSymbols, SymbolInfo},
         SymbolCatalog,
     },
 };
@@ -70,22 +72,6 @@ fn is_system_module(module: &Option<String>) -> bool {
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct AppleDebugImage {
-    pub debug_id: String,
-    pub image_addr: String,
-    #[serde(default)]
-    pub image_vmaddr: Option<String>,
-    #[serde(default)]
-    pub image_size: Option<u64>,
-    #[serde(default)]
-    pub code_file: Option<String>,
-    #[serde(default, rename = "type")]
-    pub image_type: Option<String>,
-    #[serde(default)]
-    pub arch: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct RawAppleFrame {
     pub instruction_addr: Option<String>,
     pub symbol_addr: Option<String>,
@@ -105,10 +91,10 @@ impl RawAppleFrame {
         &self,
         team_id: i32,
         catalog: &C,
-        debug_images: &[AppleDebugImage],
+        debug_images: &[DebugImage],
     ) -> Result<Vec<Frame>, UnhandledError>
     where
-        C: SymbolCatalog<OrChunkId<AppleRef>, ParsedAppleSymbols>,
+        C: SymbolCatalog<OrChunkId<AppleRef>, ParsedNativeSymbols>,
     {
         tracing::debug!(
             "[apple-debug] resolve() called: instruction_addr={:?}, module={:?}, function={:?}, debug_images_count={}",
@@ -151,10 +137,10 @@ impl RawAppleFrame {
         &self,
         team_id: i32,
         catalog: &C,
-        debug_images: &[AppleDebugImage],
+        debug_images: &[DebugImage],
     ) -> Result<Vec<Frame>, ResolveError>
     where
-        C: SymbolCatalog<OrChunkId<AppleRef>, ParsedAppleSymbols>,
+        C: SymbolCatalog<OrChunkId<AppleRef>, ParsedNativeSymbols>,
     {
         let instruction_addr = self
             .instruction_addr
@@ -163,20 +149,24 @@ impl RawAppleFrame {
                 "missing instruction_addr".into(),
             ))?;
 
-        let instruction_addr = parse_hex_address(instruction_addr)?;
+        let instruction_addr =
+            native::parse_hex_address(instruction_addr).map_err(AppleError::from)?;
         tracing::debug!(
             "[apple-debug] resolve_impl: parsed instruction_addr=0x{:x}",
             instruction_addr
         );
 
-        let debug_image = self.find_debug_image(instruction_addr, debug_images)?;
+        let debug_image =
+            native::find_debug_image(instruction_addr, self.image_addr.as_deref(), debug_images)
+                .map_err(AppleError::from)?;
         tracing::debug!(
             "[apple-debug] resolve_impl: matched debug_image debug_id={}, image_addr={}",
             debug_image.debug_id,
             debug_image.image_addr
         );
 
-        let relative_addr = self.calculate_relative_addr(instruction_addr, debug_image)?;
+        let relative_addr = native::calculate_relative_addr(instruction_addr, debug_image)
+            .map_err(AppleError::from)?;
 
         // Subtract 1 from return-address frames so the lookup targets the call instruction
         // rather than the instruction after it, giving the correct source line.
@@ -193,12 +183,12 @@ impl RawAppleFrame {
             "[apple-debug] resolve_impl: looking up symbols for chunk_id={}",
             debug_image.debug_id
         );
-        let symbols: Arc<ParsedAppleSymbols> = catalog
+        let symbols: Arc<ParsedNativeSymbols> = catalog
             .lookup(team_id, OrChunkId::chunk_id(debug_image.debug_id.clone()))
             .await?;
         tracing::debug!("[apple-debug] resolve_impl: symbols loaded successfully");
 
-        let symbol_infos = symbols.lookup(lookup_addr)?;
+        let symbol_infos = symbols.lookup(lookup_addr).map_err(AppleError::from)?;
         if symbol_infos.is_empty() {
             return Err(AppleError::SymbolNotFound(lookup_addr).into());
         }
@@ -248,60 +238,7 @@ impl RawAppleFrame {
         Ok(frames)
     }
 
-    fn find_debug_image<'a>(
-        &self,
-        instruction_addr: u64,
-        debug_images: &'a [AppleDebugImage],
-    ) -> Result<&'a AppleDebugImage, AppleError> {
-        let frame_image_addr = self
-            .image_addr
-            .as_ref()
-            .and_then(|addr| parse_hex_address(addr).ok());
-
-        for image in debug_images {
-            let image_base = parse_hex_address(&image.image_addr).ok();
-
-            if let (Some(frame_addr), Some(base)) = (frame_image_addr, image_base) {
-                if frame_addr == base {
-                    return Ok(image);
-                }
-            }
-
-            if let (Some(base), Some(size)) = (image_base, image.image_size) {
-                if instruction_addr >= base && instruction_addr < base.saturating_add(size) {
-                    return Ok(image);
-                }
-            }
-        }
-
-        Err(AppleError::NoMatchingDebugImage)
-    }
-
-    fn calculate_relative_addr(
-        &self,
-        instruction_addr: u64,
-        debug_image: &AppleDebugImage,
-    ) -> Result<u64, AppleError> {
-        let image_addr = parse_hex_address(&debug_image.image_addr)?;
-
-        if instruction_addr < image_addr {
-            return Err(AppleError::InvalidAddress(format!(
-                "instruction_addr 0x{:x} < image_addr 0x{:x}",
-                instruction_addr, image_addr
-            )));
-        }
-
-        // Calculate the offset from the runtime load address
-        // The symcache already contains addresses relative to the binary's VM base,
-        // so we just need the offset from where it was loaded
-        Ok(instruction_addr - image_addr)
-    }
-
-    fn build_resolved_frame(
-        &self,
-        symbol_info: &crate::symbol_store::apple::SymbolInfo,
-        _debug_image: &AppleDebugImage,
-    ) -> Frame {
+    fn build_resolved_frame(&self, symbol_info: &SymbolInfo, _debug_image: &DebugImage) -> Frame {
         // Override in_app to false for system frameworks or compiler-generated code
         let is_compiler_generated = symbol_info
             .filename
@@ -410,14 +347,18 @@ impl RawAppleFrame {
         frame
     }
 
-    pub fn frame_id(&self, debug_images: &[AppleDebugImage]) -> String {
+    pub fn frame_id(&self, debug_images: &[DebugImage]) -> String {
         let mut hasher = Sha512::new();
 
         // Absolute instruction addresses are ASLR-slid per process launch, so hashing
         // them directly gives the same logical frame a different id every launch and
         // the frame record cache never hits. Hash the launch-invariant
         // (debug image, relative offset) identity instead whenever we can compute it.
-        match self.launch_invariant_addr(debug_images) {
+        match native::launch_invariant_addr(
+            self.instruction_addr.as_deref(),
+            self.image_addr.as_deref(),
+            debug_images,
+        ) {
             Some((debug_id, relative_addr)) => {
                 hasher.update(debug_id.as_bytes());
                 hasher.update(format!("rel:0x{relative_addr:x}").as_bytes());
@@ -443,15 +384,6 @@ impl RawAppleFrame {
 
         format!("{:x}", hasher.finalize())
     }
-
-    fn launch_invariant_addr(&self, debug_images: &[AppleDebugImage]) -> Option<(String, u64)> {
-        let instruction_addr = parse_hex_address(self.instruction_addr.as_ref()?).ok()?;
-        let debug_image = self.find_debug_image(instruction_addr, debug_images).ok()?;
-        let relative_addr = self
-            .calculate_relative_addr(instruction_addr, debug_image)
-            .ok()?;
-        Some((debug_image.debug_id.clone(), relative_addr))
-    }
 }
 
 /// Infer the programming language from a source filename for syntax highlighting.
@@ -466,11 +398,6 @@ fn lang_from_filename(filename: Option<&str>) -> &'static str {
         Some("h") => "c", // headers default to C
         _ => "swift",     // default for Apple platforms
     }
-}
-
-fn parse_hex_address(s: &str) -> Result<u64, AppleError> {
-    let s = s.trim().trim_start_matches("0x").trim_start_matches("0X");
-    u64::from_str_radix(s, 16).map_err(|_| AppleError::InvalidAddress(s.to_string()))
 }
 
 impl From<&RawAppleFrame> for Frame {
@@ -505,189 +432,6 @@ impl From<&RawAppleFrame> for Frame {
 #[cfg(test)]
 mod test {
     use super::*;
-
-    #[test]
-    fn test_parse_hex_address_with_0x_prefix() {
-        assert_eq!(parse_hex_address("0x100000000").unwrap(), 0x100000000);
-        assert_eq!(parse_hex_address("0X100000000").unwrap(), 0x100000000);
-    }
-
-    #[test]
-    fn test_parse_hex_address_without_prefix() {
-        assert_eq!(parse_hex_address("100000000").unwrap(), 0x100000000);
-        assert_eq!(parse_hex_address("deadbeef").unwrap(), 0xdeadbeef);
-    }
-
-    #[test]
-    fn test_parse_hex_address_with_whitespace() {
-        assert_eq!(parse_hex_address("  0x100000000  ").unwrap(), 0x100000000);
-    }
-
-    #[test]
-    fn test_parse_hex_address_invalid() {
-        assert!(parse_hex_address("not_hex").is_err());
-        assert!(parse_hex_address("0xGGGG").is_err());
-    }
-
-    #[test]
-    fn test_calculate_relative_addr_with_vmaddr() {
-        let frame = RawAppleFrame {
-            instruction_addr: Some("0x100004000".to_string()),
-            symbol_addr: None,
-            image_addr: Some("0x100000000".to_string()),
-            image_uuid: None,
-            module: None,
-            function: None,
-            filename: None,
-            lineno: None,
-            colno: None,
-            meta: CommonFrameMetadata::default(),
-        };
-
-        let debug_image = AppleDebugImage {
-            debug_id: "test-uuid".to_string(),
-            image_addr: "0x100000000".to_string(),
-            image_vmaddr: Some("0x100000000".to_string()),
-            image_size: Some(0x10000),
-            code_file: None,
-            image_type: None,
-            arch: None,
-        };
-
-        let result = frame
-            .calculate_relative_addr(0x100004000, &debug_image)
-            .unwrap();
-        assert_eq!(result, 0x4000);
-    }
-
-    #[test]
-    fn test_calculate_relative_addr_default_vmaddr() {
-        let frame = RawAppleFrame {
-            instruction_addr: Some("0x100004000".to_string()),
-            symbol_addr: None,
-            image_addr: Some("0x100000000".to_string()),
-            image_uuid: None,
-            module: None,
-            function: None,
-            filename: None,
-            lineno: None,
-            colno: None,
-            meta: CommonFrameMetadata::default(),
-        };
-
-        let debug_image = AppleDebugImage {
-            debug_id: "test-uuid".to_string(),
-            image_addr: "0x100000000".to_string(),
-            image_vmaddr: None,
-            image_size: Some(0x10000),
-            code_file: None,
-            image_type: None,
-            arch: None,
-        };
-
-        let result = frame
-            .calculate_relative_addr(0x100004000, &debug_image)
-            .unwrap();
-        assert_eq!(result, 0x4000);
-    }
-
-    #[test]
-    fn test_find_debug_image_by_image_addr() {
-        let frame = RawAppleFrame {
-            instruction_addr: Some("0x100004000".to_string()),
-            symbol_addr: None,
-            image_addr: Some("0x100000000".to_string()),
-            image_uuid: None,
-            module: None,
-            function: None,
-            filename: None,
-            lineno: None,
-            colno: None,
-            meta: CommonFrameMetadata::default(),
-        };
-
-        let debug_images = vec![
-            AppleDebugImage {
-                debug_id: "other-uuid".to_string(),
-                image_addr: "0x200000000".to_string(),
-                image_vmaddr: None,
-                image_size: Some(0x10000),
-                code_file: None,
-                image_type: None,
-                arch: None,
-            },
-            AppleDebugImage {
-                debug_id: "matching-uuid".to_string(),
-                image_addr: "0x100000000".to_string(),
-                image_vmaddr: None,
-                image_size: Some(0x10000),
-                code_file: None,
-                image_type: None,
-                arch: None,
-            },
-        ];
-
-        let result = frame.find_debug_image(0x100004000, &debug_images).unwrap();
-        assert_eq!(result.debug_id, "matching-uuid");
-    }
-
-    #[test]
-    fn test_find_debug_image_by_address_range() {
-        let frame = RawAppleFrame {
-            instruction_addr: Some("0x100004000".to_string()),
-            symbol_addr: None,
-            image_addr: None, // No image_addr on frame
-            image_uuid: None,
-            module: None,
-            function: None,
-            filename: None,
-            lineno: None,
-            colno: None,
-            meta: CommonFrameMetadata::default(),
-        };
-
-        let debug_images = vec![AppleDebugImage {
-            debug_id: "range-match".to_string(),
-            image_addr: "0x100000000".to_string(),
-            image_vmaddr: None,
-            image_size: Some(0x10000),
-            code_file: None,
-            image_type: None,
-            arch: None,
-        }];
-
-        let result = frame.find_debug_image(0x100004000, &debug_images).unwrap();
-        assert_eq!(result.debug_id, "range-match");
-    }
-
-    #[test]
-    fn test_find_debug_image_no_match() {
-        let frame = RawAppleFrame {
-            instruction_addr: Some("0x300000000".to_string()),
-            symbol_addr: None,
-            image_addr: None,
-            image_uuid: None,
-            module: None,
-            function: None,
-            filename: None,
-            lineno: None,
-            colno: None,
-            meta: CommonFrameMetadata::default(),
-        };
-
-        let debug_images = vec![AppleDebugImage {
-            debug_id: "some-uuid".to_string(),
-            image_addr: "0x100000000".to_string(),
-            image_vmaddr: None,
-            image_size: Some(0x10000),
-            code_file: None,
-            image_type: None,
-            arch: None,
-        }];
-
-        let result = frame.find_debug_image(0x300000000, &debug_images);
-        assert!(matches!(result, Err(AppleError::NoMatchingDebugImage)));
-    }
 
     #[sqlx::test(migrations = "./tests/test_migrations")]
     async fn test_apple_symbolication(db: sqlx::PgPool) {
@@ -787,7 +531,7 @@ mod test {
             meta: CommonFrameMetadata::default(),
         };
 
-        let debug_images = vec![AppleDebugImage {
+        let debug_images = vec![DebugImage {
             debug_id: chunk_id.clone(),
             image_addr: "0x100000000".to_string(),
             image_vmaddr: Some("0x100000000".to_string()),
@@ -943,7 +687,7 @@ mod test {
             meta: CommonFrameMetadata::default(),
         };
 
-        let debug_images = vec![AppleDebugImage {
+        let debug_images = vec![DebugImage {
             debug_id: chunk_id.clone(),
             image_addr: "0x100000000".to_string(),
             image_vmaddr: Some("0x100000000".to_string()),
@@ -1020,8 +764,8 @@ mod test {
         }
     }
 
-    fn image_at(debug_id: &str, image_addr: u64) -> AppleDebugImage {
-        AppleDebugImage {
+    fn image_at(debug_id: &str, image_addr: u64) -> DebugImage {
+        DebugImage {
             debug_id: debug_id.to_string(),
             image_addr: format!("0x{image_addr:x}"),
             image_vmaddr: None,
