+++ rust/cymbal/src/stages/resolution/frame.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::{
     error::UnhandledError,
     frames::{Frame, RawFrame},
-    langs::apple::AppleDebugImage,
+    langs::native::DebugImage,
     metric_consts::FRAME_RESOLVER_OPERATOR,
     stages::{pipeline::HandledError, resolution::ResolutionStage},
     types::{
@@ -21,7 +21,7 @@ impl FrameResolver {
     pub async fn resolve_exception_list_frames(
         team_id: i32,
         list: ExceptionList,
-        debug_images: Arc<Vec<AppleDebugImage>>,
+        debug_images: Arc<Vec<DebugImage>>,
         ctx: ResolutionStage,
     ) -> Result<ExceptionList, UnhandledError> {
         let res = Batch::from(list.0)
@@ -42,7 +42,7 @@ impl FrameResolver {
     pub async fn resolve_exception_frames(
         team_id: i32,
         mut exc: Exception,
-        debug_images: &[AppleDebugImage],
+        debug_images: &[DebugImage],
         ctx: ResolutionStage,
     ) -> Result<Exception, UnhandledError> {
         exc.stack = match exc.stack {
@@ -72,7 +72,7 @@ impl FrameResolver {
     pub async fn resolve_frame(
         team_id: i32,
         frame: &RawFrame,
-        debug_images: &[AppleDebugImage],
+        debug_images: &[DebugImage],
         ctx: ResolutionStage,
     ) -> Result<Vec<Frame>, UnhandledError> {
         let _permit = ctx.acquire_symbol_resolution_permit().await?;
